ar-06: fix CI regressions from the format attributes
macOS/clang: off_t is long long there (long on Linux), so the raw %ld for
file_stat.st_size in toml_parser tripped -Werror=format under the new format
attribute. Cast to long, matching config.c's identical size check.

flawfinder: the 'printf' token inside the format(printf,...) attribute macro is
flagged as a level-4 format hit by CI's flawfinder version; annotate the line
with a flawfinder: ignore directive (it is the compiler archetype, not a call).

+++ src/toml_parser.c
@@ -75,8 +75,11 @@ int toml_parse_file(const char *file_path, toml_document_t *doc) {
     
     /* Security: Check file size limit */
     if (file_stat.st_size > TOML_MAX_FILE_SIZE) {
-        set_error(ERR_CONFIG_INVALID, "Configuration file too large: %ld bytes (max: %d)", 
-                  file_stat.st_size, TOML_MAX_FILE_SIZE);
+        /* Cast to long: off_t is long long on macOS/clang (Wformat error under
+         * -Werror) but long on Linux; the value is bounded small here (just
+         * over the max), so no truncation. Mirrors config.c's size check. */
+        set_error(ERR_CONFIG_INVALID, "Configuration file too large: %ld bytes (max: %d)",
+                  (long)file_stat.st_size, TOML_MAX_FILE_SIZE);
         return -1;
     }
